Jacobian is working I think!
The important thing is to make the blocks of the Jacobian periodic in the x direction by appropriately boinking some diagonals into places. It is a bit of a hack rn, so I think that I should add lots of tests. Alos, the hack that I have used is v suboptimal performance wise.